web/terraguard-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.20
psycopg2-binary==2.9.10
psycopg[binary]==3.2.4
opencv-python
ultralytics
numpy
easyocr
httpx
